+++ packages/survey-creator-core/src/components/question.ts
@@ -15,7 +15,11 @@ import { DragDropHelper } from "survey-core";
 import { getLocString } from "../editorLocalization";
 import { QuestionConverter } from "../questionconverter";
 import { IPortableDragEvent, IPortableMouseEvent } from "../utils/events";
-import { isPropertyVisible, propertyExists } from "../utils/utils";
+import {
+  isPropertyVisible,
+  propertyExists,
+  toggleHovered
+} from "../utils/utils";
 import { ActionContainerViewModel } from "./action-container-view-model";
 import "./question.scss";
 
@@ -28,7 +32,6 @@ export class QuestionAdornerViewModel extends ActionContainerViewModel<SurveyMod
     public templateData: SurveyTemplateRendererTemplateData
   ) {
     super(creator, surveyElement);
-    this.actionContainer.setItems(this.getContextActions());
   }
   select(model: QuestionAdornerViewModel, event: IPortableMouseEvent) {
     if (!model.surveyElement.isInteractiveDesignElement) {
@@ -36,7 +39,7 @@ export class QuestionAdornerViewModel extends ActionContainerViewModel<SurveyMod
     }
     event.stopPropagation();
     event.cancelBubble = true;
-    model.creator.selectElement(model.surveyElement);
+    model.creator.selectElement(model.surveyElement, undefined, false);
     return true;
   }
   css() {
@@ -58,6 +61,23 @@ export class QuestionAdornerViewModel extends ActionContainerViewModel<SurveyMod
   }
   get isDraggable() {
     return true;
+  }
+  public hover(event: MouseEvent, element: HTMLElement) {
+    if (!this.surveyElement.isInteractiveDesignElement) {
+      return;
+    }
+    toggleHovered(event, element);
+  }
+  protected updateElementAllowOptions(options: any, operationsAllow: boolean) {
+    super.updateElementAllowOptions(options, operationsAllow);
+    this.updateActionVisibility(
+      "convertTo",
+      operationsAllow && options.allowChangeType
+    );
+    this.updateActionVisibility(
+      "isrequired",
+      operationsAllow && options.allowChangeRequired
+    );
   }
 
   public get isEmptyElement(): boolean {
@@ -175,32 +195,20 @@ export class QuestionAdornerViewModel extends ActionContainerViewModel<SurveyMod
     return requiredAction;
   }
 
-  protected getContextActions(): Array<Action> {
-    if (this.creator.readOnly) return [];
-
-    const items = super.getContextActions();
-
+  protected buildActions(items: Array<Action>) {
+    super.buildActions(items);
     let element = this.surveyElement;
-    let opts: any = element["allowingOptions"];
-    if (!opts) opts = {};
-
-    if (opts.allowChangeType === undefined || opts.allowChangeType) {
-      if (!element.isPanel) {
-        items.push(this.createConverToAction());
-      }
+    if (!element.isPanel) {
+      items.push(this.createConverToAction());
     }
 
     if (
-      (opts.allowChangeRequired === undefined || opts.allowChangeRequired) &&
       typeof element["isRequired"] !== "undefined" &&
       propertyExists(element, "isRequired") &&
       isPropertyVisible(element, "isRequired")
     ) {
       items.push(this.createRequiredAction());
     }
-
-    this.creator.onElementMenuItemsChanged(element, items);
-    return items;
   }
   protected duplicate() {
     var newElement = this.creator.fastCopyQuestion(this.surveyElement);