+++ packages/survey-creator-core/src/settings.ts
@@ -38,19 +38,33 @@ export var settings = {
    */
   questionConvertMode: QuestionConvertMode.AllTypes,
   propertyGrid: {
-    useButtonGroup: true
+    useButtonGroup: true,
+    showNavigationButtons: true
   },
   /**
    * Notification settings
    */
-  notifications : {
+  notifications: {
     lifetime: 2000
   },
   /**
    * Auto save parameters
    */
-   autoSave: {
+  autoSave: {
     delay: 500
+  },
+  /**
+   * Drag Drop Settings
+   */
+  dragDrop: {
+    restrictDragQuestionBetweenPages: false
+  },
+  /**
+   * Creator layout settings
+   */
+  layout: {
+    showTabs: true,
+    showToolbar: true
   }
 
   //TODO add maximumColumnCount