+++ packages/survey-creator-react/src/components/tabs/Designer.tsx
@@ -7,9 +7,8 @@ import {
 } from "survey-react-ui";
 import { CreatorBase, TabDesignerViewModel } from "@survey/creator";
 import { CreatorSurveyPageComponent } from "../Page";
-import { SurveyCreatorToolbox } from "../toolbox/Toolbox";
 import { SurveyPageNavigator } from "../page-navigator/PageNavigator";
-import PropertyGridComponent from "../../PropertyGrid";
+import { SurveyNavigation } from "../Navigation";
 
 interface ITabDesignerComponentProps {
   data: TabDesignerViewModel<SurveyModel>;
@@ -22,15 +21,14 @@ export class TabDesignerComponent extends SurveyElementBase<
   private get model(): TabDesignerViewModel<SurveyModel> {
     return this.props.data;
   }
+  protected get creator(): CreatorBase<SurveyModel> {
+    return this.model.creator;
+  }
   protected getStateElement(): Base {
     return this.model;
   }
-  renderElement(): JSX.Element {
-    const creator: CreatorBase<SurveyModel> = this.model.creator;
-    const survey: SurveyModel = creator.survey;
-    const designerTabClassName = "svc-tab-designer " + survey.css.root;
-
-    const surveyPages: JSX.Element[] = survey.pages.map((page: PageModel) => {
+  protected renderPages(): JSX.Element[] {
+    const surveyPages: JSX.Element[] = this.creator.survey.pages.map((page: PageModel) => {
       return (
         <div
           className={"svc-page"}
@@ -38,9 +36,9 @@ export class TabDesignerComponent extends SurveyElementBase<
           key={page.id}
         >
           <CreatorSurveyPageComponent
-            survey={survey}
+            survey={this.creator.survey}
             page={page}
-            creator={creator}
+            creator={this.creator}
           ></CreatorSurveyPageComponent>
         </div>
       );
@@ -54,103 +52,51 @@ export class TabDesignerComponent extends SurveyElementBase<
           key={this.model.newPage.id}
         >
           <CreatorSurveyPageComponent
-            survey={survey}
+            survey={this.creator.survey}
             page={this.model.newPage}
-            creator={creator}
+            creator={this.creator}
           ></CreatorSurveyPageComponent>
         </div>
       );
     }
+    return surveyPages;
+  }
+  renderElement(): JSX.Element {
+    const survey: SurveyModel = this.creator.survey;
+    const designerTabClassName = "svc-tab-designer " + survey.css.root;
 
     return (
       <React.Fragment>
         <div className="svc-flex-column">
-          {ReactElementFactory.Instance.createElement('svc-toolbox', { creator: creator })}
+          {ReactElementFactory.Instance.createElement('svc-toolbox', { creator: this.creator })}
         </div>
-        <div className={"svc-tab-designer " + creator.survey.css.root}>
-          <div className={creator.survey.css.container}>
+        <div className={"svc-tab-designer " + survey.css.root}>
+          <div className={survey.css.container}>
             <div
               onClick={function () {
-                creator.selectElement(creator.survey);
+                this.creator.selectElement(survey);
               }}
             >
               <SurveyHeader survey={survey}></SurveyHeader>
             </div>
-            <DesignerSurveyNavigationBlock
-              survey={creator.survey}
+            <SurveyNavigation
+              survey={survey}
               location="top"
             />
-            {surveyPages}
-            <DesignerSurveyNavigationBlock
-              survey={creator.survey}
+            {this.renderPages()}
+            <SurveyNavigation
+              survey={survey}
               location="bottom"
-              css={creator.survey.css}
+              css={survey.css}
             />
           </div>
         </div>
         <SurveyPageNavigator
-          creator={creator}
-          pages={creator.pagesController.pages}
+          creator={this.creator}
+          pages={this.creator.pagesController.pages}
         ></SurveyPageNavigator>
-          {ReactElementFactory.Instance.createElement('svc-property-grid', { model: creator })}
+          {ReactElementFactory.Instance.createElement('svc-property-grid', { model: this.creator })}
       </React.Fragment>
-    );
-  }
-}
-export class DesignerSurveyNavigationBlock extends SurveyElementBase<any, any> {
-  componentDidMount() {
-    this.setHandler();
-  }
-  componentDidUpdate(prevProps: any, prevState: any) {
-    this.setHandler();
-  }
-  private setHandler() {
-    if (
-      !this.survey ||
-      this.survey.onPropertyChanged.hasFunc(this.onPropChangedHandler)
-    )
-      return;
-    this.survey.onPropertyChanged.add(this.onPropChangedHandler);
-  }
-  private onPropChangedHandler = (sender: any, options: any): any => {
-    if (this.isRendering) return;
-    const reactiveProps = [
-      "showProgressBar",
-      "progressBarType",
-      "currentPageValue"
-    ];
-    if (reactiveProps.indexOf(options.name) < 0) return;
-    var val: any = {};
-    for (var i = 0; i < reactiveProps.length; i++) {
-      var propName = reactiveProps[i];
-      val[propName] = this.survey[propName];
-    }
-    this.setState(val);
-  };
-  componentWillUnmount() {
-    if (this.survey) {
-      this.survey.onPropertyChanged.remove(this.onPropChangedHandler);
-    }
-  }
-
-  protected get survey(): SurveyModel {
-    return this.props.survey;
-  }
-  protected get location(): string {
-    return this.props.location;
-  }
-  protected get isTop(): boolean {
-    return this.location == "top";
-  }
-  protected canRender(): boolean {
-    return this.isTop
-      ? this.survey.isShowProgressBarOnTop
-      : this.survey.isShowProgressBarOnBottom;
-  }
-  renderElement(): JSX.Element {
-    return ReactElementFactory.Instance.createElement(
-      "sv-progress-" + this.survey.progressBarType.toLowerCase(),
-      { survey: this.survey, css: this.survey.css, isTop: this.isTop }
     );
   }
 }