+++ packages/survey-creator/src/components/toolbar.ts
@@ -1,11 +1,11 @@
 import * as ko from "knockout";
-import { IActionBarItem } from "survey-knockout";
+import { IAction } from "survey-knockout";
 
 import "./toolbar.scss";
 var template = require("./toolbar.html");
 
 export class Toolbar {
-  constructor(public items: ko.ObservableArray<IActionBarItem> | Array<IActionBarItem>) { }
+  constructor(public items: ko.ObservableArray<IAction> | Array<IAction>) {}
   get hasItems() {
     return (ko.unwrap(this.items) || []).length > 0;
   }
@@ -15,7 +15,7 @@ ko.components.register("svd-toolbar", {
   viewModel: {
     createViewModel: (params, componentInfo) => {
       return new Toolbar(params.items);
-    }
+    },
   },
-  template: template
+  template: template,
 });