+++ packages/survey-creator/src/entries/index.ts
@@ -112,6 +112,7 @@ import "../localization/russian";
 import "../localization/tajik";
 import "../localization/dutch";
 import "../localization/croatian";
+import "../localization/finnish";
 
 export {
   Translation,