+++ packages/survey-creator-react/CHANGELOG.md
@@ -2,4 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.8.58-alpha.1](https://github.com/surveyjs/survey-creator/compare/v1.8.58...v1.8.58-alpha.0) (2021-07-28)
+
+### [1.8.57-alpha.1](https://github.com/surveyjs/survey-creator/compare/v1.8.57...v1.8.57-alpha.1) (2021-07-20)
+
 ### [0.0.1-alpha.0](https://github.com/surveyjs/survey-creator/compare/v1.8.49...v0.0.1-alpha.0) (2021-05-27)